Move filter classes and functions to filters.py
